Migrated BaseIconExtension to New Impl
